cpufreq: added Intelliminmax CPU governor
A rewrite of the old Min Max governor and has 3 cpu states: Idle, UI and Max.
Intelliminmax (intellimm) governor is designed to work with the newer SOCs
with fixed voltage rails (ie MSM89XX+ SOCs). It is designed to work within
those fixed voltage ranges in order to maximize battery performance while
creating a smooth UI operations. It is battery friendly and spends most of
the time at lower frequencies.