cpufreq: added Preservative CPU governor
Preservative is based on the idea that the CPU will consume a lot of power
when it changes frequency. It is based on the conservative governor.
The idea is that it will stay at the step specified (702MHz selected
by the creator Bedalus) unless needed. You will notice it will hover
around 702 a lot, and not go above too much, and only to min freq
when NOTHING is happening at all. This is most beneficial when you
are doing something like reading; the screen is static or playing
light games that won't need boosting any more
The governor comes from Moob kernel for nexus 4