cpufreq: added Wheatley CPU governor
This governor is build on “ondemand” but increases the C4 (the sleep state)
state time of the CPU and doing so trying to save juice. So the results show
that Wheatley works as intended and ensures that the C4 state is used whenever
the task allows a proper efficient usage of the C4 state. For more demanding
tasks which cause a large number of wakeups and prevent the efficient usage of
the C4 state, the governor resorts to the next best power saving mechanism and
scales down the frequency. So with the new highly-flexible Wheatley governor
one can have the best of both worlds. Obviously, this governor is only available
on multi-core devices.

Wheatley is a more performance orientated governor as it scales more aggressively
than ondemand and sticks with higher frequencies.